Count reads in a library like HTSeq

This function creates a granges object from the txdb and calls countReadsInRanges. Uses for this function may be during QC to generate log2cpm over the CDS regions in protein coding transcriptome, for example

countLibrary(
  bam_path,
  annote_db,
  feature_type,
  strandedness,
  num_threads = 1,
  single_end_reads_flag = TRUE,
  mapq_filter = 20L
)

Arguments

bam_path
path to a bam file. the index must also exist

annote_db
txdb object. see makeTxDbFromGFF

feature_type
one of c("exon", "cds"). Determines what feature to count over

strandedness
strandedness of library determined by protocol. One of c("reverse", "same", "unstranded") reverse means the counted reads will be on the opposite strand of the feature, same means they will be on the same strand. unstranded counts reads over the feature regardless of strand

num_threads
number of threads available for parallelization. Default to 1

single_end_reads_flag
whether or not the library is single or paired. Note: this is not written to deal with paired end reads currently, though you can try

mapq_filter
quality filter on read alignment. Default to 20, same as HTSeq. see https://htseq.readthedocs.io/en/release_0.11.1/count.html

Value

A RangedSummarizedExperiment. See summarizeOverlaps

See also

Examples

if (FALSE) library(brentlabRnaSeqTools)
library(AnnotationDbi)

#> Loading required package: stats4

#> Loading required package: BiocGenerics

#> Loading required package: parallel

#> 
#> Attaching package: 'BiocGenerics'

#> The following objects are masked from 'package:parallel':
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB

#> The following objects are masked from 'package:stats':
#> 
#>     IQR, mad, sd, var, xtabs

#> The following objects are masked from 'package:base':
#> 
#>     anyDuplicated, append, as.data.frame, basename, cbind, colnames,
#>     dirname, do.call, duplicated, eval, evalq, Filter, Find, get, grep,
#>     grepl, intersect, is.unsorted, lapply, Map, mapply, match, mget,
#>     order, paste, pmax, pmax.int, pmin, pmin.int, Position, rank,
#>     rbind, Reduce, rownames, sapply, setdiff, sort, table, tapply,
#>     union, unique, unsplit, which.max, which.min

#> Loading required package: Biobase

#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.

#> Loading required package: IRanges

#> Loading required package: S4Vectors

#> 
#> Attaching package: 'S4Vectors'

#> The following objects are masked from 'package:base':
#> 
#>     expand.grid, I, unname

library(tidyverse)

#> ── Attaching packages ─────────────────────────────────────── tidyverse 1.3.1 ──

#> ✔ ggplot2 3.3.5     ✔ dplyr   1.0.7
#> ✔ tidyr   1.1.3     ✔ stringr 1.4.0
#> ✔ readr   2.0.0     ✔ forcats 0.5.1
#> ✔ purrr   0.3.4     

#> ── Conflicts ────────────────────────────────────────── tidyverse_conflicts() ──
#> ✖ dplyr::collapse()   masks IRanges::collapse()
#> ✖ dplyr::combine()    masks Biobase::combine(), BiocGenerics::combine()
#> ✖ dplyr::desc()       masks IRanges::desc()
#> ✖ tidyr::expand()     masks S4Vectors::expand()
#> ✖ dplyr::filter()     masks stats::filter()
#> ✖ dplyr::first()      masks S4Vectors::first()
#> ✖ dplyr::lag()        masks stats::lag()
#> ✖ ggplot2::Position() masks BiocGenerics::Position(), base::Position()
#> ✖ purrr::reduce()     masks IRanges::reduce()
#> ✖ dplyr::rename()     masks S4Vectors::rename()
#> ✖ dplyr::select()     masks AnnotationDbi::select()
#> ✖ dplyr::slice()      masks IRanges::slice()

kn99_db = loadDb("data/kn99_db.sqlite")

#> Error in dbFileConnect(file): DB file 'data/kn99_db.sqlite' not found

combined_df = getMetadata(database_info$kn99_host,
                          database_info$kn99_db_name,
                          Sys.getenv("db_username"),
                          Sys.getenv("password"))

#> Error: fe_sendauth: no password supplied

run_5102 = combined_df %>%
 filter(runNumber == 5102,
        experimentDesign == '90minuteInduction',
        genotype1 == "CNAG_00000")

#> Error in filter(., runNumber == 5102, experimentDesign == "90minuteInduction",     genotype1 == "CNAG_00000"): object 'combined_df' not found

wt_fastq = run_5102[[1, 'fastqFileName']]

#> Error in eval(expr, envir, enclos): object 'run_5102' not found

bam_path = createBamPath(5102,
                         wt_fastq,
                         align_expr_prefix =
                         "/mnt/htcf_scratch/chasem/rnaseq_pipeline/align_count_results")

#> Error in basename(fastq_filename): object 'wt_fastq' not found

cds_count = countLibrary(bam_path, kn99_db, 'cds', 'reverse', 8)

#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'cdsBy': object 'kn99_db' not found

cds_counts_df = as_tibble(assay(cds_counts), rownames = "gene_name")

#> Error in assay(cds_counts): could not find function "assay"