Create a rank-response data frame

Create a rank-response data frame

Usage

sort_rank_mean_expr(
  expression_df,
  expression_src,
  binding_df,
  binding_src,
  lfc_thres,
  padj_thres,
  rank_resolution = 10,
  num_rows = 15
)

Arguments

expression_df

a three column dataframe with the column names c('gene', 'log2FoldChange', 'padj') which describes expression data. It may have more columns, but those three must exist

expression_src

the source of the data, eg 'kemmeren' or 'brentlab'

binding_df

a two column dataframe dataframe with the columns c('gene', 'binding_signal'). It may have more columns, but those must exist

binding_src

the source of the data, eg 'chip-chip' or 'calling-cards'

lfc_thres

expression log2foldchange threshold. default is 0

padj_thres

expression padj threshold. default is .05

rank_resolution

how to bin ranks. Default is 10, which means that ranks are binned into groups of 10 (eg the first group are the 10 highest)

num_rows

the number of rows to return. Default is 15. For instance, if the rank_resolution is 10, and you only want to look at the first 150 ranked genes, then num_rows is 15.

Value

a rank_response dataframe